+++ packages/contracts/src/core/permission/PermissionConditionUpgradeable.sol
@@ -10,6 +10,7 @@ import {IPermissionCondition} from "./IPermissionCondition.sol";
 /// @title PermissionConditionUpgradeable
 /// @author Aragon Association - 2023
 /// @notice An abstract contract for upgradeable or cloneable contracts to inherit from and to support customary permissions depending on arbitrary on-chain state.
+/// @custom:security-contact sirt@aragon.org
 abstract contract PermissionConditionUpgradeable is
     ERC165Upgradeable,
     IPermissionCondition,