+++ packages/contracts/src/core/plugin/Plugin.sol
@@ -12,6 +12,7 @@ import {IPlugin} from "./IPlugin.sol";
 /// @title Plugin
 /// @author Aragon Association - 2022-2023
 /// @notice An abstract, non-upgradeable contract to inherit from when creating a plugin being deployed via the `new` keyword.
+/// @custom:security-contact sirt@aragon.org
 abstract contract Plugin is IPlugin, ERC165, DaoAuthorizable, ProtocolVersion {
     /// @notice Constructs the plugin by storing the associated DAO.
     /// @param _dao The DAO contract.