+++ packages/contracts/src/core/plugin/PluginUUPSUpgradeable.sol
@@ -14,6 +14,7 @@ import {IPlugin} from "./IPlugin.sol";
 /// @title PluginUUPSUpgradeable
 /// @author Aragon Association - 2022-2023
 /// @notice An abstract, upgradeable contract to inherit from when creating a plugin being deployed via the UUPS pattern (see [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822)).
+/// @custom:security-contact sirt@aragon.org
 abstract contract PluginUUPSUpgradeable is
     IPlugin,
     ERC165Upgradeable,