+++ packages/contracts/src/core/plugin/dao-authorizable/DaoAuthorizableUpgradeable.sol
@@ -11,6 +11,7 @@ import {_auth} from "../../utils/auth.sol";
 /// @author Aragon Association - 2022-2023
 /// @notice An abstract contract providing a meta-transaction compatible modifier for upgradeable or cloneable contracts to authorize function calls through an associated DAO.
 /// @dev Make sure to call `__DaoAuthorizableUpgradeable_init` during initialization of the inheriting contract.
+/// @custom:security-contact sirt@aragon.org
 abstract contract DaoAuthorizableUpgradeable is ContextUpgradeable {
     /// @notice The associated DAO managing the permissions of inheriting contracts.
     IDAO private dao_;