+++ packages/contracts/src/framework/dao/DAORegistry.sol
@@ -11,6 +11,7 @@ import {isSubdomainValid} from "../utils/RegistryUtils.sol";
 /// @title Register your unique DAO subdomain
 /// @author Aragon Association - 2022-2023
 /// @notice This contract provides the possibility to register a DAO.
+/// @custom:security-contact sirt@aragon.org
 contract DAORegistry is InterfaceBasedRegistry, ProtocolVersion {
     /// @notice The ID of the permission required to call the `register` function.
     bytes32 public constant REGISTER_DAO_PERMISSION_ID = keccak256("REGISTER_DAO_PERMISSION");