+++ packages/contracts/src/framework/plugin/setup/PluginSetup.sol
@@ -14,6 +14,7 @@ import {IPluginSetup} from "./IPluginSetup.sol";
 /// @title PluginSetup
 /// @author Aragon Association - 2022-2023
 /// @notice An abstract contract that developers have to inherit from to write the setup of a plugin.
+/// @custom:security-contact sirt@aragon.org
 abstract contract PluginSetup is ERC165, IPluginSetup, ProtocolVersion {
     /// @inheritdoc IPluginSetup
     function prepareUpdate(