+++ packages/contracts/src/framework/plugin/setup/PluginSetupProcessor.sol
@@ -21,6 +21,7 @@ import {PluginSetupRef, hashHelpers, hashPermissions, _getPreparedSetupId, _getA
 /// @author Aragon Association - 2022-2023
 /// @notice This contract processes the preparation and application of plugin setups (installation, update, uninstallation) on behalf of a requesting DAO.
 /// @dev This contract is temporarily granted the `ROOT_PERMISSION_ID` permission on the applying DAO and therefore is highly security critical.
+/// @custom:security-contact sirt@aragon.org
 contract PluginSetupProcessor is ProtocolVersion {
     using ERC165Checker for address;