+++ packages/contracts/src/token/ERC20/governance/GovernanceERC20.sol
@@ -17,6 +17,7 @@ import {IERC20MintableUpgradeable} from "../IERC20MintableUpgradeable.sol";
 /// @title GovernanceERC20
 /// @author Aragon Association
 /// @notice An [OpenZeppelin `Votes`](https://docs.openzeppelin.com/contracts/4.x/api/governance#Votes) compatible [ERC-20](https://eips.ethereum.org/EIPS/eip-20) token that can be used for voting and is managed by a DAO.
+/// @custom:security-contact sirt@aragon.org
 contract GovernanceERC20 is
     IERC20MintableUpgradeable,
     Initializable,