+++ packages/contracts/src/token/ERC20/governance/GovernanceWrappedERC20.sol
@@ -24,6 +24,7 @@ import {IGovernanceWrappedERC20} from "./IGovernanceWrappedERC20.sol";
 /// 2. call `depositFor` to wrap them, which safely transfers the underlying [ERC-20](https://eips.ethereum.org/EIPS/eip-20) tokens to the contract and mints wrapped [ERC-20](https://eips.ethereum.org/EIPS/eip-20) tokens.
 /// To get the [ERC-20](https://eips.ethereum.org/EIPS/eip-20) tokens back, the owner of the wrapped tokens can call `withdrawFor`, which  burns the wrapped [ERC-20](https://eips.ethereum.org/EIPS/eip-20) tokens and safely transfers the underlying tokens back to the owner.
 /// @dev This contract intentionally has no public mint functionality because this is the responsibility of the underlying [ERC-20](https://eips.ethereum.org/EIPS/eip-20) token contract.
+/// @custom:security-contact sirt@aragon.org
 contract GovernanceWrappedERC20 is
     IGovernanceWrappedERC20,
     Initializable,